Update alacritty installer

+++ alacritty/install.sh
@@ -30,14 +30,18 @@ if ! command -v alacritty &>/dev/null; then
 	fi
 fi
 
+## System Configuration
+### Natural Scrolling, tapping
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+	sudo rm -rf /usr/share/X11/xorg.conf.d/40-libinput.conf
+	sudo ln -sf "${DOTFILE_DIRECTORY}/alacritty/40-libinput.conf" "/usr/share/X11/xorg.conf.d/40-libinput.conf"
+fi
+
 ## Link dotfiles
 rm -rf "${HOME}/.config/alacritty"
-sudo rm -rf /usr/share/X11/xorg.conf.d/40-libinput.conf
 mkdir -p "${HOME}/.config/alacritty"
 ln -sf "${DOTFILE_DIRECTORY}/alacritty/alacritty.yml" "${HOME}/.config/alacritty/alacritty.yml"
 ln -sf "${DOTFILE_DIRECTORY}/alacritty/gruvbox_material.yml" "${HOME}/.config/alacritty/gruvbox_material.yml"
-### Natural Scrolling, tapping
-sudo ln -sf "${DOTFILE_DIRECTORY}/alacritty/40-libinput.conf" "/usr/share/X11/xorg.conf.d/40-libinput.conf"
 
 ## ChromeOS and Ubuntu Desktop Icon
 if [[ "$OSTYPE" == "linux-gnu"* ]]; then
